backend/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy  
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg
google-cloud-bigquery
flask-cors